restructure and add maven
