+++ src/main/java/com/hotelpassmanagement/Template.java
@@ -1,0 +1,7 @@
+package com.hotelpassmanagement;
+class Template {
+    public static void main(String args[]) {
+        Login l = new Login();
+        l.loginpage();
+    }
+}
