+++ src/main/java/com/hotelpassmanagement/runner/StuRunner.java
@@ -1,0 +1,10 @@
+package com.hotelpassmanagement.runner;
+import com.hotelpassmanagement.model.Student;
+
+class StuRunner {
+    public static void main(String args[]) {
+        Student stu = new Student();
+        stu.SetStudent();
+        stu.show();
+    }
+}
